initial refactoring effort
